Code stuff
If you're going to trace through the source code, I forgot to add comments to the TeamMatchup class. It's basically a double for-loop that runs in O(n^2) time theoretically speaking (to be specific, it always runs at O(108) since the numbers are always fixed to be 6 and 18). The first for-loop is used to cycle through each member of a team of 6 Pokemon. The second for-loop cycles through the type match-ups of each Pokemon and tallies the total numbers of damage immunities, damage resistances (both single and double), weaknesses (both single and double), and neutral damage (neither resistant nor weak to) that the entire team has.